Feat: Finshed with Weather api
