+++ server/routes/welcome.ts
@@ -6,7 +6,7 @@ const router = express.Router()
 // GET /api/v1/welcome/
 router.get('/', (req, res) => {
   try {
-    res.json({ statement: 'Welcome to external APIs!' } as Welcome)
+    res.json({ statement: 'Welcome to API Playground!' } as Welcome)
   } catch (err) {
     if (err instanceof Error) {
       res.status(500).send((err as Error).message)
